Removed a line in default.project.json that was referencing a folder which didnt exist, causing the sourcemap to never be able to generate. Hopefully its not somehow important...?
Cleaned up and fully commented ServerScriptService/Managers/SaveManager,
making it more readable with more descriptive variable names
(v, i -> skinname, skin; and so on. Felt important for a template
project), and more indentation and spacing for lines at my personal
descretion (modify as you please– I just find it easier to work with
this way).

Potentially fixed SaveManager corrupting data by having it attempt
reading/writing from and to the datastore multiple times (with a cap on
attempts) until it finally succeeds. Datastores are too unreliable to
be one-and-done, so this is common practice– and we were having issues
on our end with save data corruption, so hopefully this fixes it...?