Stablize Utils, Inventory, and SaveManager w/ non-existent characters
Fixed absentee killers causing an error in both Utils.GetCharacterModule
and StarterPlayerScripts/UI/Inventory. One with an error-handling line
trying to print out a skin variable (which is optional, and errors if
not present because you cant concatenate a string with something that
doesn't exist, lol), and one with a simple check made before trying to
load a character card with a nonexsistent module.

Fixed the entire game grinding to a halt if the default equipped
killer/survivor in the template playerdata points to one that doesn't
exist. Now the game does a sanity check on those two values, and if they
point to a character that doesn't exist, it'll select the first one it
can grab from the killer or survivor folders respectively, as well as
print out a line pointing towards those valueobjects for a developer
to change to whom they wish.

    This does not kick in if the value is manually changed to an
    existing killer/survivor, so it will never override the developer's
    choice (unless if they delete their own killer/survivor without
    changing the default, for some reason).