Prep actions-status-discord, need input from Dyscarn to finalize
Prepared a more customizable github action to update the ds-commits
discord channel, which incl. a way to disable sending the webhook on
things like comments and to only send commits or pull requests, for
example. It requires some setup on Dyscarn's part so I need to wait for
them to finalize it.